Fix epan-sys bindgen due to missing type
Fixes "Unable to get layout information?" error when building epan-sys
with the bindgen feature enabled. This error occurs because the union
"wtap_pseudo_header" is used in epan/packet_info.h (included by
epan/packet.h) but the definition is not included. The union is defined
in wiretap/wtap.h which we can include to fix the error